refactor: move logic to domain + add default timespan values
